vision_manager/requirements.txt

streamlit
reportlab
psycopg2-binary
